Changed the generator to a config based system
Added a config file system to the generator so that any folder can have a ".uigencfg" file that sources the template file, defines the output file format, and loop startIndex and count parameters.

These changes allow for the bracket substitution to do a single arithmetic operation on 'i'. Allowed operations are: '+' add, '-' subtract, '*' multiply, '/' divide, and '%' modulo.